requirements.txt
Django==1.5.4
futures==3.0.5
geojson==1.3.2
mercantile==0.8.3
Pillow==3.2.0
psycopg2==2.6.1
requests==2.10.0
requests-futures==0.9.7
Shapely==1.5.15
six==1.10.0
